ovirt-guest-tools-iso: Shorten CD Volume Identifier
Create CD iso image with volume identifier of oVirt-WGT-$VERSION
instead of oVirt-Tools-$VERSION.

With a CD created with joliet extensions, vol id is effectively limited
to 16 chars. This change allows to include there the full version for
VERSION of X.Y.Z where X and Y are a single digit each and Z is up to
two digits.

Change-Id: I75a9d8eba16a5c7956a7510dcac12ebd86cf6bcf
Signed-off-by: Yedidyah Bar David <didi@redhat.com>